% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk_generator.R
\name{Bulk_generator}
\alias{Bulk_generator}
\title{Pseudo Bulk generator from single cell data}
\usage{
Bulk_generator(
  sce,
  phenoData,
  Num.mixtures = 1000,
  nb_CT_random = TRUE,
  pool.size = 100,
  min.percentage = 1,
  max.percentage = 99,
  seed = 24,
  forbidden_CT = NULL,
  showDensity = FALSE
)
}
\arguments{
\item{sce}{numeric matrix or data.frame containing the single cell data. columns : cells, rows : genes}

\item{phenoData}{data.frame containing cellID and cellType as character}

\item{Num.mixtures}{integer number of pseudo bulk to generate (default = 1000)}

\item{nb_CT_random}{whether the cells types present in pseudo bulk should be selected randomly or not. If not, all the cell types will be kept (default = TRUE)}

\item{pool.size}{integer number of single cell per mixture (default = 100)}

\item{min.percentage}{minimum percentage of cells to put in a mixture (if cell type is selceted) (default = 1)}

\item{max.percentage}{maximum percentage of cells to put in a mixture (if cell type is selceted) (default = 99)}

\item{seed}{random seed (default = 24)}

\item{forbidden_CT}{character vector indicating cellType not to put in the mixture.}

\item{showDensity}{whether plot an overview of the cell type's density along mixtures (default = FALSE)}
}
\value{
list with T a data.frame of the Num.mixture pseudo bulk and P a data.frame of the cellType proportions used to build the mixtures.
}
\description{
Pseudo Bulk generator from single cell data
}
\examples{
Bulk_generator(sc_counts, phenoData)
}
